fix of ability functionality
updating the properties from the ability classes did not seem to work but invoke methods in PlayerMovement that do will work. Also the Time.timeScale needs to be called in the main Thread so we just set a bool and in update we fix it if the bool is true.